basket 기능 추가

+++ src/main/java/com/thing/item/dto/ItemDetailResponseDTO.java
@@ -12,10 +12,10 @@ public class ItemDetailResponseDTO {
 
     private ClientInfoDTO ownerInfo;
     private Item item;
-    private Long basketCount;
+    private Integer basketCount;
     private Boolean isLike;
 
-    public static ItemDetailResponseDTO from(ClientInfoDTO clientInfoDTO, Item item, Long basketCount, Boolean isLike){
+    public static ItemDetailResponseDTO from(ClientInfoDTO clientInfoDTO, Item item, Integer basketCount, Boolean isLike){
         return new ItemDetailResponseDTO(
                 clientInfoDTO,
                 item,
